Support NDEF text records beyond 127 characters
The keyboard capped text input at 127 chars (NFC_TOOLS_NDEF_BUF1_SIZE - 1),
even though the NDEF encoder, page writer and tag-capacity gate already
handle long messages. Raise the write input buffer and the read-side
decoded value buffer from 128 to 1024 bytes, and widen two uint8_t
text-length variables in the keyboard to size_t so backspace and cursor
handling stay correct past 255 chars. Bump version to 1.3.

+++ views/keyboard/nfc_tools_keyboard.c
@@ -235,7 +235,7 @@ static void kb_draw_page(Canvas* canvas, uint8_t x, uint8_t y, bool sel, uint8_t
 
 static void kb_draw_callback(Canvas* canvas, void* _model) {
     KbModel*    model        = _model;
-    uint8_t     text_length  = model->text_buffer ? (uint8_t)strlen(model->text_buffer) : 0;
+    size_t      text_length  = model->text_buffer ? strlen(model->text_buffer) : 0;
     uint8_t     needed_width = (uint8_t)(canvas_width(canvas) - 8);
     uint8_t     start_pos    = 4;
     const char* text         = model->text_buffer;
@@ -313,7 +313,7 @@ static void kb_draw_callback(Canvas* canvas, void* _model) {
 
 static void kb_backspace(KbModel* model) {
     // If default text is selected, treat it as a single unit to delete
-    uint8_t len = model->clear_default_text ? 1 : (uint8_t)strlen(model->text_buffer);
+    size_t len = model->clear_default_text ? 1 : strlen(model->text_buffer);
     if(len > 0) model->text_buffer[len - 1] = '\0';
 }
